+++ database/migrations/2026_01_20_122815_create_municipios_table.php
@@ -11,12 +11,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('municipios', function (Blueprint $table) {
-            $table->id();
-            $table->string('municipio')->unique();
-            $table->foreignId('departamento_id')->constrained('departamentos')->onDelete('restrict');
-            $table->timestamps();
-        });
+        if (! Schema::hasTable('municipios')) {
+            Schema::create('municipios', function (Blueprint $table) {
+                $table->id();
+                $table->string('municipio')->unique();
+                $table->foreignId('departamento_id')->constrained('departamentos')->onDelete('restrict');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
